lab56

+++ src/main/java/com/example/vector/processes/ParamOutput.java
@@ -2,23 +2,22 @@ package com.example.vector.processes;
 
 public class ParamOutput {
 
-    private final int ProjectionOnX;
-
-    private final int ProjectionOnY;
+    private final double ProjectionOnX;
+    private final double ProjectionOnY;
     private final double vectorRate;
 
 
     public ParamOutput(
-            int ProjectionOnX,
-            int ProjectionOnY,
+            double ProjectionOnX,
+            double ProjectionOnY,
             double vectorRate
     ) {
         this.ProjectionOnX = ProjectionOnX;
         this.ProjectionOnY = ProjectionOnY;
         this.vectorRate = vectorRate;
     }
-    public int getProjectionOnX() {return ProjectionOnX; }
-    public int getProjectionOnY() {return ProjectionOnY; }
+    public double getProjectionOnX() {return ProjectionOnX; }
+    public double getProjectionOnY() {return ProjectionOnY; }
 
     public double getVectorRate() {return vectorRate; }
 }
